--- WeatherAPI/Services/IAirQualityPollenService.cs
@@ -1,11 +1,0 @@
-﻿using WeatherAPI.DTOs;
-using WeatherAPI.Models;
-
-namespace WeatherAPI.Services
-{
-    public interface IAirQualityPollenService
-    {
-        Task<GetPollenDTO> GetPollenData(string cityName);
-        Task<PollenSuggestion> GetPollenSuggestion(string cityName);
-    }
-}
