+++ WeatherAPI/Services/IAirQualityService.cs
@@ -1,9 +1,13 @@
 ﻿using WeatherAPI.DTOs;
+using WeatherAPI.Models;
 
 namespace WeatherAPI.Services
 {
-    public interface IAirQualityParticulateMatterService
+    public interface IAirQualityService
     {
+        Task<GetPollenDTO> GetPollenData(string cityName);
+        Task<PollenSuggestion> GetPollenSuggestion(string cityName);
+
         Task<GetAirQualityParticulateMatterResponseDTO> GetAirQualityParticulateMatterByCityName(string cityName);
 
         Task<SuggestionsOnAirQualityParticulateMatterDTO> SuggestionsOnAirQualityParticulateMatterByCityName(string cityName);
